docker/requirements.txt
altair==5.4.1
asgiref==3.8.1
async-timeout==5.0.1
attrs==24.2.0
blinker==1.9.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
contourpy==1.3.1
cycler==0.12.1
Django==5.1.3
django-debug-toolbar==4.4.6
django-redis==5.4.0
djangorestframework==3.15.2
dnspython==2.7.0
fonttools==4.55.0
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.1.1
idna==3.10
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
mdurl==0.1.2
narwhals==1.14.1
numpy==2.1.3
packaging==24.2
pandas==2.2.3
pillow==11.0.0
plotly==5.24.1
protobuf==5.28.3
psycopg2-binary==2.9.10
pyarrow==18.0.0
pydeck==0.9.1
Pygments==2.18.0
pymongo==4.10.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
redis==5.2.0
referencing==0.35.1
requests==2.32.3
rich==13.9.4
rpds-py==0.21.0
six==1.16.0
smmap==5.0.1
SQLAlchemy==2.0.36
sqlparse==0.5.2
streamlit==1.40.1
tenacity==9.0.0
toml==0.10.2
tornado==6.4.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
watchdog==6.0.0
